Moving over the new Dashboard to replace the old
refs: https://github.com/TryGhost/Team/issues/1631
