+++ app/components/dashboard/charts/anchor.js
