+++ app/components/dashboard/charts/engagement.js
@@ -1,6 +1,6 @@
 import Component from '@glimmer/component';
 import {action} from '@ember/object';
-import {formatNumber} from '../../../../helpers/format-number';
+import {formatNumber} from '../../../helpers/format-number';
 import {inject as service} from '@ember/service';
 import {tracked} from '@glimmer/tracking';
