+++ app/components/dashboard/charts/overview.hbs
@@ -2,7 +2,7 @@
     <article {{did-insert this.loadCharts}} class="gh-dashboard5-box is-secondary">
         <div class="gh-dashboard5-columns">
             <div class="gh-dashboard5-column">
-                <Dashboard::v5::Parts::Metric
+                <Dashboard::Parts::Metric
                     @label={{gh-pluralize this.totalMembers "Total member" without-count=true}}
                     @value={{this.totalMembersFormatted}}
                     @trends={{this.hasTrends}}
@@ -10,7 +10,7 @@
                     @large={{true}} /> 
             </div>
             <div class="gh-dashboard5-column">
-                <Dashboard::v5::Parts::Metric
+                <Dashboard::Parts::Metric
                     @label={{gh-pluralize this.paidMembers "Paid member" without-count=true}}
                     @value={{this.paidMembersFormatted}}
                     @trends={{this.hasTrends}}
@@ -18,7 +18,7 @@
                     @large={{true}} /> 
             </div>
             <div class="gh-dashboard5-column">
-                <Dashboard::v5::Parts::Metric
+                <Dashboard::Parts::Metric
                     @label={{gh-pluralize this.freeMembers "Free member" without-count=true}}
                     @value={{this.freeMembersFormatted}}
                     @trends={{this.hasTrends}}
