+++ app/components/dashboard/charts/overview.js
@@ -1,6 +1,6 @@
 import Component from '@glimmer/component';
 import {action} from '@ember/object';
-import {formatNumber} from '../../../../helpers/format-number';
+import {formatNumber} from '../../../helpers/format-number';
 import {inject as service} from '@ember/service';
 
 export default class Overview extends Component {
