+++ app/components/dashboard/charts/paid-breakdown.hbs
@@ -1,7 +1,7 @@
 <div class="gh-dashboard5-minichart gh-dashboard5-breakdown">
     <div class="gh-dashboard5-content">
         <div class="gh-dashboard5-data">
-            <Dashboard::v5::Parts::Metric
+            <Dashboard::Parts::Metric
                 @label={{this.chartTitle}} />
 
             <div class="gh-dashboard5-legend">
