+++ app/components/dashboard/charts/paid-breakdown.js
