+++ app/components/dashboard/charts/paid-mix.hbs
@@ -1,7 +1,7 @@
 <div class="gh-dashboard5-minichart gh-dashboard5-mix {{if this.isChartTiers 'is-tiers'}}">
     <div class="gh-dashboard5-content">
         <div class="gh-dashboard5-data">
-            <Dashboard::v5::Parts::Metric
+            <Dashboard::Parts::Metric
                 @label="Paid mix" /> 
 
             {{#if this.isChartCadence}}
