+++ app/components/dashboard/charts/paid-mix.js
