+++ app/components/dashboard/charts/paid-mrr.hbs
@@ -1,7 +1,7 @@
 <div class="gh-dashboard5-minichart gh-dashboard5-mrr">
     <div class="gh-dashboard5-content">
         <div class="gh-dashboard5-data">
-            <Dashboard::v5::Parts::Metric
+            <Dashboard::Parts::Metric
                 @label={{this.chartTitle}}
                 @value="{{this.currentMRRFormatted}}"
                 @trends={{this.hasTrends}}
