+++ app/components/dashboard/charts/paid-mrr.js
@@ -3,7 +3,7 @@
 import Component from '@glimmer/component';
 import moment from 'moment';
 import {getSymbol} from 'ghost-admin/utils/currency';
-import {ghPriceAmount} from '../../../../helpers/gh-price-amount';
+import {ghPriceAmount} from '../../../helpers/gh-price-amount';
 import {inject as service} from '@ember/service';
 
 const DATE_FORMAT = 'D MMM, YYYY';
