+++ app/components/dashboard/parts/percentage.hbs
