+++ app/components/dashboard/prototype/control-panel.hbs
