+++ app/components/dashboard/prototype/control-panel.js
