+++ app/components/dashboard/resources/community.hbs
