+++ app/components/dashboard/resources/newsletter.hbs
