+++ app/components/dashboard/resources/newsletter.js
