+++ app/components/dashboard/resources/resources.hbs
