+++ app/components/dashboard/resources/resources.js
