+++ app/components/dashboard/resources/staff-picks.hbs
