+++ app/components/dashboard/resources/staff-picks.js
