+++ app/components/dashboard/resources/whats-new.hbs
