+++ app/styles/app-dark.css
@@ -69,7 +69,6 @@
 @import "layouts/fullscreen-wizard.css";
 @import "layouts/post-preview.css";
 @import "layouts/dashboard.css";
-@import "layouts/dashboard-v5.css";
 @import "layouts/tiers.css";
 @import "layouts/offers.css";
